Add bin/setup shell script
This shell script should allow you to more easily set up new dev
environments in the future. I've commented in the file exactly what it
does and it can be run from the command line from your dotfiles
directory by typing `bin/setup.sh` and hitting `Enter/Return`.

It should be easy to follow the patter to add additional default
packages and settings should you desire.

I also changed a setting in your vimrc so that it works with vim and
neovim